+++ python/python/tests/test_transaction.py
@@ -17,6 +17,7 @@ from glide.async_commands.sorted_set import (
     LexBoundary,
     RangeByIndex,
     ScoreBoundary,
+    ScoreFilter,
 )
 from glide.async_commands.transaction import (
     BaseTransaction,
@@ -50,6 +51,7 @@ async def transaction_test(
     key13 = "{{{}}}:{}".format(keyslot, get_random_string(3))  # sorted set
     key14 = "{{{}}}:{}".format(keyslot, get_random_string(3))  # sorted set
     key15 = "{{{}}}:{}".format(keyslot, get_random_string(3))  # sorted set
+    key16 = "{{{}}}:{}".format(keyslot, get_random_string(3))  # sorted set
 
     value = datetime.now(timezone.utc).strftime("%m/%d/%Y, %H:%M:%S")
     value2 = get_random_string(5)
@@ -201,8 +203,12 @@ async def transaction_test(
     args.append("bar")
     transaction.sadd(key7, ["foo", "bar"])
     args.append(2)
+    transaction.sunionstore(key7, [key7, key7])
+    args.append(2)
     transaction.spop_count(key7, 4)
     args.append({"foo", "bar"})
+    transaction.smove(key7, key7, "non_existing_member")
+    args.append(False)
 
     transaction.zadd(key8, {"one": 1, "two": 2, "three": 3, "four": 4})
     args.append(4)
@@ -252,6 +258,11 @@ async def transaction_test(
     args.append(0)
     transaction.zdiffstore(key8, [key8, key8])
     args.append(0)
+    if not await check_if_server_version_lt(redis_client, "7.0.0"):
+        transaction.zmpop([key8], ScoreFilter.MAX)
+        args.append(None)
+        transaction.zmpop([key8], ScoreFilter.MAX, 1)
+        args.append(None)
 
     transaction.zadd(key13, {"one": 1.0, "two": 2.0})
     args.append(2)
@@ -259,6 +270,12 @@ async def transaction_test(
     args.append(["one", "two"])
     transaction.zdiff_withscores([key13, key8])
     args.append({"one": 1.0, "two": 2.0})
+    if not await check_if_server_version_lt(redis_client, "7.0.0"):
+        transaction.zintercard([key13, key8])
+        args.append(0)
+        transaction.zintercard([key13, key8], 1)
+        args.append(0)
+
     transaction.zadd(key14, {"one": 1, "two": 2})
     args.append(2)
     transaction.zadd(key15, {"one": 1.0, "two": 2.0, "three": 3.5})
@@ -300,6 +317,16 @@ async def transaction_test(
     args.append("0-2")
     transaction.xtrim(key11, TrimByMinId(threshold="0-2", exact=True))
     args.append(1)
+
+    min_version = "7.0.0"
+    if not await check_if_server_version_lt(redis_client, min_version):
+        transaction.zadd(key16, {"a": 1, "b": 2, "c": 3, "d": 4})
+        args.append(4)
+        transaction.bzmpop([key16], ScoreFilter.MAX, 0.1)
+        args.append([key16, {"d": 4.0}])
+        transaction.bzmpop([key16], ScoreFilter.MIN, 0.1, 2)
+        args.append([key16, {"a": 1.0, "b": 2.0}])
+
     return args